+++ src/schema.rs
@@ -176,6 +176,14 @@ impl SchemaType for AccountAddress {
 impl SchemaType for ContractAddress {
     fn get_type() -> Type { Type::ContractAddress }
 }
+impl SchemaType for Address {
+    fn get_type() -> Type {
+        Type::Enum(Vec::from([
+            (String::from("account"), Fields::Unnamed(Vec::from([Type::AccountAddress]))),
+            (String::from("contract"), Fields::Unnamed(Vec::from([Type::ContractAddress]))),
+        ]))
+    }
+}
 impl SchemaType for Timestamp {
     fn get_type() -> Type { Type::Timestamp }
 }