+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased changes
 - Add public validation functions to contract and receive names
 - Add new cases for NewContractNameError and NewReceiveNameError
+- Implement `SchemaType` for `Address`.
+- Derive `PartialOrd` and `Ord` for `ContractAddress` and `Address`.
 
 ## concordium-contracts-common 0.4.0 (2021-05-12)